fix(archive): drop useless u64 conversion in disk-space preflight
nix::sys::statvfs::Statvfs::blocks_available() already returns u64 on
this platform; the u64::from() wrapper was a no-op clippy first caught
running on real Linux.

+++ crates/substrate-archive/src/resource_limit.rs
@@ -67,7 +67,7 @@ fn check_disk_space_sync(path: &Path, required_bytes: u64) -> SubstrateResult<()
         correlation_id: Some(uuid::Uuid::now_v7()),
     })?;
 
-    let available: u64 = u64::from(stat.blocks_available()) * stat.block_size();
+    let available: u64 = stat.blocks_available() * stat.block_size();
     let needed = required_bytes.saturating_add(FREE_SPACE_CUSHION_BYTES);
 
     if available < needed {
